Use MSI @ 16 MHz not 24 MHz
This clock speed reduction eliminates occasional hard faults when
debugging in stop modes.  Seems that maybe the wake from stop modes
is too fast for the flash when the debugger keeps clocks running in
stop modes.  Adding just one more wait state was enough to eliminate
the hard faults, but I didn't want to do that.  A clock-speed reduction
is suitable for this demo app.

Note for posterity that tracking the problem was difficult.  Any code
changes invalidate the test.  I finally got a set of code that caused
the hard fault reliably, then I used the debugger to add a wait state.
The problem went away reliably with the extra wait state, and the
problem returned reliably without it.  I did verify that the SYSCLK
keeps running in stop mode, so no overshoot concerns.  I also verified
that the problem occurs *only* with DBGMCU_CR_DBG_STOP set.  I also
verified STOP 0 experiences the same problem as the other STOP modes,
again doing so carefully with the debugger choosing the target stop mode
and not a code change.  (STOP 0 keeps the main regulator running.)